Gaus works

+++ src/main/java/ru/nsu/icg/lab2/model/transformations/FilterApplicator.java
@@ -36,21 +36,21 @@ public class FilterApplicator extends Transformation {
         for (int y = borderStep; y < height - borderStep; y++){
             for (int x = borderStep; x < width - borderStep; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = index + inY * width + inX;
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -59,21 +59,21 @@ public class FilterApplicator extends Transformation {
         for (int y = 0; y < borderStep; y++){
             for (int x = 0; x < borderStep; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = Math.abs(inY + y) * width + Math.abs(inX + x);
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -81,21 +81,21 @@ public class FilterApplicator extends Transformation {
         for (int y = 0; y < borderStep; y++){
             for (int x = borderStep; x < width - borderStep; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = Math.abs(inY + y) * width + inX + x;
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -103,21 +103,21 @@ public class FilterApplicator extends Transformation {
         for (int y = 0; y < borderStep; y++){
             for (int x = width - borderStep; x < width; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = Math.abs(inY + y) * width + ((inX + x) % width);
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -125,21 +125,21 @@ public class FilterApplicator extends Transformation {
         for (int y = borderStep; y < height - borderStep; y++){
             for (int x = width - borderStep; x < width; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = (y + inY) * width + ((inX + x) % width);
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -147,21 +147,21 @@ public class FilterApplicator extends Transformation {
         for (int y = height - borderStep; y < height; y++){
             for (int x = width - borderStep; x < width; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = ((y + inY) % height)  * width + ((inX + x) % width);
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -169,21 +169,21 @@ public class FilterApplicator extends Transformation {
         for (int y = height - borderStep; y < height; y++){
             for (int x = borderStep; x < width - borderStep; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = ((y + inY) % height)  * width + inX + x;
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -191,21 +191,21 @@ public class FilterApplicator extends Transformation {
         for (int y = height - borderStep; y < height; y++){
             for (int x = 0; x < borderStep; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = ((y + inY) % height)  * width + Math.abs(inX + x);
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue);
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
 
@@ -213,21 +213,21 @@ public class FilterApplicator extends Transformation {
         for (int y = borderStep; y < height - borderStep; y++){
             for (int x = 0; x < borderStep; x++){
                 int index = y * width + x;
-                int red = 0;
-                int green = 0;
-                int blue = 0;
+                double red = 0;
+                double green = 0;
+                double blue = 0;
                 for (int inX = - borderStep; inX <= borderStep; inX++){
                     for (int inY = -borderStep; inY <= borderStep; inY++){
                         int inXIndex = inX + borderStep;
                         int inYIndex = inY + borderStep;
                         int gridIndex = (y + inY) * width + Math.abs(inX + x);
                         int pixelColor = oldGrid[gridIndex];
-                        red += (int) ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
-                        green += (int) ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
-                        blue += (int) ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
+                        red += ((((pixelColor & 0x00FF0000) >> 16) * matrix[inYIndex][inXIndex]));
+                        green += ((((pixelColor & 0x0000FF00) >> 8) * matrix[inYIndex][inXIndex]));
+                        blue += ((((pixelColor & 0x000000FF)) * matrix[inYIndex][inXIndex]));
                     }
                 }
-                newGrid[index] |= counter.color(red, green, blue) ;
+                newGrid[index] |= counter.color((int) red,(int) green,(int) blue);
             }
         }
